some more tests and stuff

+++ positional-macro/src/lib.rs
@@ -13,7 +13,7 @@ pub fn peano(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
 }
 
 fn peano_impl(input: LitInt) -> syn::Result<TokenStream> {
-    let int = input.base10_parse::<u32>()?;
+    let int = input.base10_parse::<u16>()?;
     let mut res = quote! { ::positional::peano::Zero };
 
     for _ in 0..int {
